backend/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
requests
opencv-python-headless
eventlet
qrcode
Pillow